[export] Add ability to run eagerly on UnflattenedModule (#133996)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/133996

Added the contextmanager, `_disable_interpreter`, which is meant to put around a call to `unflatten`. This will generate an UnflattendModule and sub-InterpreterModules which will not use torch.fx.Interpreter to run eagerly. We want to have this as a state of the module instead of a contextmanager around running the module because it's not clear where we are calling the unflattened module.

This seems to improve the performance: https://fb.workplace.com/groups/1075192433118967/posts/1473590629945810/?comment_id=1473621763276030

Test Plan: CI

Reviewed By: pianpwk

Differential Revision: D60939034